requirements.txt
openai>=1.40.0
pillow>=10.0.0
